migrate pillar unit tests to pytest
